refactor: restructure loaders
