+++ example/loaders/blood_sample.py
@@ -1,4 +1,4 @@
-import psycopmlutils.loaders as loaders
+import psycopmlutils.loaders.raw as loaders
 
 if __name__ == "__main__":
     df = loaders.load_lab_results.LoadLabResults.hdl()
