+++ example/loaders/load_physical_visits.py
@@ -1,4 +1,4 @@
-from psycopmlutils.loaders import LoadVisits
+from psycopmlutils.loaders.raw import LoadVisits
 
 if __name__ == "__main__":
     df = LoadVisits.physical_visits_to_psychiatry()
