+++ example/loaders/load_sql.py
@@ -1,4 +1,4 @@
-from psycopmlutils.loaders.sql_load import sql_load
+from psycopmlutils.loaders.raw.sql_load import sql_load
 
 if __name__ == "__main__":
     view = "[psycop_t2d_train]"
