--- src/psycopmlutils/loaders/__init__.py
@@ -1,8 +1,0 @@
-from .load_demographic import LoadDemographic  # noqa
-from .load_diagnoses import LoadDiagnoses  # noqa
-from .load_ids import LoadIDs  # noqa
-from .load_lab_results import LoadLabResults  # noqa
-from .load_medications import LoadMedications  # noqa
-from .load_outcomes import LoadOutcome  # noqa
-from .load_visits import LoadVisits  # noqa
-from .sql_load import sql_load  # noqa
