+++ src/psycopmlutils/loaders/flattened/__init__.py
@@ -1,0 +1,1 @@
+from .local_feature_loaders import load_split_predictors_and_outcomes  # noqa
